update to report number of critical hits
To avoid contention why some rolls might be higher than the
max damage dice, the results of the roll also include the number
of critical hits.

The console logs are still there for verification, but it should
be more clearer.

This also includes a version bump and updates to report the
version in the HTML.